Don't trust unauthenticated embedded revocation data
Code review (P1): xades:RevocationValues live in UnsignedSignatureProperties
and are not covered by the signature, so short-circuiting the revocation
verdict to "good" from them let a tampered container pass a revoked signer
with no network. Authenticating them offline would require verifying the
OCSP/CRL signature and responder chain against a trusted issuer plus
freshness, which isn't generally possible offline (the issuer cert may not
be in the container).

- Remove the embedded-revocation short-circuit; the verdict now comes only
  from live OCSP/CRL. When revocation can't be checked it is INDETERMINATE,
  never "good"
- Drop RevocationCheckOptions.embeddedOCSP/embeddedCRL/atTime and
  RevocationResult.fromEmbedded; delete the embedded evaluator
- Still parse and expose SignatureInfo.revocationValues as RAW, explicitly
  unvalidated material for consumers doing their own LTV validation
- Add a regression test: offline revocation is indeterminate (not pass) while
  document/signature/certificate/timestamp/issuer checks still verify offline

The original INDETERMINATE-on-large-CRL bug stays fixed via the parseCRL
node-cap fallback on the live path (verified: live CRL returns "good").
